1. Removing Annotations for Simple Names @DisplayName. 2. Change the return type of the involved method from int to long. 3. Using isIn instead of containsAnyOf for correct assertions. 4. Directly throw InterruptedException without capturing it. 5. Latch calls await method with timeout. 6. Merge comment lines. 7. Use the get method to wait for the future to complete, just like the original Groovy code. 8. In the mongoSpan method parameters, use SpanDataAsset instead of trace and index. 9. In the mongoSpan method, optimizing assertion error prompts.

+++ instrumentation/mongo/mongo-4.0/javaagent/src/test/java/io/opentelemetry/javaagent/instrumentation/mongo/v4_0/MongoClientTest.java
@@ -81,9 +81,9 @@ class MongoClientTest extends AbstractMongoClientTest<MongoCollection<Document>>
   }
 
   @Override
-  public int getCollection(String dbName, String collectionName) {
+  public long getCollection(String dbName, String collectionName) {
     MongoDatabase db = client.getDatabase(dbName);
-    return Math.toIntExact(db.getCollection(collectionName).estimatedDocumentCount());
+    return db.getCollection(collectionName).estimatedDocumentCount();
   }
 
   @Override
@@ -102,9 +102,9 @@ class MongoClientTest extends AbstractMongoClientTest<MongoCollection<Document>>
   }
 
   @Override
-  public int insert(MongoCollection<Document> collection) {
+  public long insert(MongoCollection<Document> collection) {
     collection.insertOne(new Document("password", "SECRET"));
-    return Math.toIntExact(collection.estimatedDocumentCount());
+    return collection.estimatedDocumentCount();
   }
 
   @Override
@@ -125,13 +125,13 @@ class MongoClientTest extends AbstractMongoClientTest<MongoCollection<Document>>
   }
 
   @Override
-  public int update(MongoCollection<Document> collection) {
+  public long update(MongoCollection<Document> collection) {
     UpdateResult result =
         collection.updateOne(
             new BsonDocument("password", new BsonString("OLDPW")),
             new BsonDocument("$set", new BsonDocument("password", new BsonString("NEWPW"))));
     collection.estimatedDocumentCount();
-    return Math.toIntExact(result.getModifiedCount());
+    return result.getModifiedCount();
   }
 
   @Override
@@ -152,11 +152,11 @@ class MongoClientTest extends AbstractMongoClientTest<MongoCollection<Document>>
   }
 
   @Override
-  public int delete(MongoCollection<Document> collection) {
+  public long delete(MongoCollection<Document> collection) {
     DeleteResult result =
         collection.deleteOne(new BsonDocument("password", new BsonString("SECRET")));
     collection.estimatedDocumentCount();
-    return Math.toIntExact(result.getDeletedCount());
+    return result.getDeletedCount();
   }
 
   @Override
